+++ src/MonoTorrent.Client/MonoTorrent.Client.csproj
@@ -112,7 +112,7 @@ Notable features include:
     
   <ItemGroup>
     <ProjectReference Include="..\MonoTorrent.BEncoding\MonoTorrent.BEncoding.csproj" />
-    <ProjectReference Include="..\MonoTorrent.Messages\MonoTorrent.PeerMessages.csproj" />
+    <ProjectReference Include="..\MonoTorrent.Messages\MonoTorrent.Messages.csproj" />
     <ProjectReference Include="..\MonoTorrent.PiecePicking\MonoTorrent.PiecePicking.csproj" />
     <ProjectReference Include="..\MonoTorrent.PieceWriter\MonoTorrent.PieceWriter.csproj" />
     <ProjectReference Include="..\MonoTorrent\MonoTorrent.csproj" />
