+++ src/MonoTorrent.sln
@@ -39,7 +39,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MonoTorrent.PieceWriter", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests.MonoTorrent.PieceWriter", "Tests\Tests.MonoTorrent.PieceWriter\Tests.MonoTorrent.PieceWriter.csproj", "{BAE86E51-93A0-499D-A4ED-3519DEB5E4DF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MonoTorrent.PeerMessages", "MonoTorrent.Messages\MonoTorrent.PeerMessages.csproj", "{79271733-DC12-4DB1-8A34-3843EEBE9DD8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MonoTorrent.Messages", "MonoTorrent.Messages\MonoTorrent.Messages.csproj", "{79271733-DC12-4DB1-8A34-3843EEBE9DD8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests.MonoTorrent.Messages", "Tests\Tests.MonoTorrent.Messages\Tests.MonoTorrent.Messages.csproj", "{A8C22AB5-9A76-4258-881E-4AF8E72CC2D5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -103,6 +105,10 @@ Global
 		{79271733-DC12-4DB1-8A34-3843EEBE9DD8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{79271733-DC12-4DB1-8A34-3843EEBE9DD8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{79271733-DC12-4DB1-8A34-3843EEBE9DD8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A8C22AB5-9A76-4258-881E-4AF8E72CC2D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A8C22AB5-9A76-4258-881E-4AF8E72CC2D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A8C22AB5-9A76-4258-881E-4AF8E72CC2D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A8C22AB5-9A76-4258-881E-4AF8E72CC2D5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -116,6 +122,7 @@ Global
 		{08D34B26-FDF5-4AFA-B0A3-2475CCA6EB7E} = {8790FB34-68AA-4A1A-A339-0E1DAAB3E180}
 		{8CE2C458-D0A9-428F-A749-4949AA113EEC} = {8790FB34-68AA-4A1A-A339-0E1DAAB3E180}
 		{BAE86E51-93A0-499D-A4ED-3519DEB5E4DF} = {8790FB34-68AA-4A1A-A339-0E1DAAB3E180}
+		{A8C22AB5-9A76-4258-881E-4AF8E72CC2D5} = {8790FB34-68AA-4A1A-A339-0E1DAAB3E180}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {FDCCD363-7E21-4AA3-9373-5DCB46A67AD1}
